feedbackgoals.component.ts:
found that the model used for directorArr is wrong.  Will need to revisit this item to remove the error showing up in the IDE.  This error is just because the types are wrong and the datafield -- despite existing on the objects in the array -- does not exist on the model the variable was typed as.  The code works, it just gives a funky red line to a file that I want to remove for clarity reason.

Changed code to access the _id field of the directorArr object.  This identifies the directors for the routing in expressjs

admin.js-routing
Added mongoose import to handle the id objects
added a console log for debugging.  will remove this soon

ToDo: Tackle the admin selecting an admin / year where data already exists. Want to recall this data and update the display to show the information back to the admin for editing.